Create Introduction to Data Engineering
This course came to me via email and I could not pass it up. I am enjoying it so far. It is interesting as my position has been changing from data science to engineering and I was not aware of the differences. This course really cleared up my understanding of the two positions.